requirementsSDXL.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
annotated-types==0.7.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///croot/attrs_1695717823297/work
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cityscapesScripts==2.2.3
coloredlogs==15.0.1
comm @ file:///croot/comm_1709322850197/work
contourpy @ file:///croot/contourpy_1700583582875/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
datasets @ file:///croot/datasets_1716911606380/work
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
docstring-parser==0.15
dominate==2.9.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.16.0
fonttools @ file:///croot/fonttools_1713551344105/work
frozenlist @ file:///croot/frozenlist_1698702560391/work
fsspec==2024.6.1
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
imageio==2.35.1
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
instructor==0.6.8
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1726064227985/work
jedi @ file:///croot/jedi_1721058342488/work
Jinja2==3.1.4
jiter==0.5.0
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
kiwisolver @ file:///work/ci_py311/kiwisolver_1676827230232/work
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mpmath==1.3.0
multidict @ file:///croot/multidict_1701096859099/work
multiprocess @ file:///croot/multiprocess_1692294385131/work
nano==0.10.0
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.3
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=5f96f274d410a1682519282ae769c877d32fdbf171aa8badec7bf5e1d3a1748a
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.47.0
opencv-python==4.10.0.84
packaging @ file:///croot/packaging_1720101850331/work
pandas @ file:///croot/pandas_1718308974269/work/dist/pandas-2.2.2-cp311-cp311-linux_x86_64.whl#sha256=3c7ce50f9f519c785bd4cdb28a0ca71f85a541f3d27b25aa9da770f953e7f2e9
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///croot/pillow_1721059439630/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
ply==3.11
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==5.28.2
psutil @ file:///work/ci_py311_2/psutil_1679337388738/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow @ file:///croot/pyarrow_1721664224170/work/python
pydantic==2.9.2
pydantic_core==2.23.4
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///croot/pyparsing_1725041624202/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
pyquaternion==0.9.9
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytorch-fid==0.3.0
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
rich==13.8.1
safetensors==0.4.5
scikit-image==0.24.0
scikit-posthocs==0.11.2
scipy==1.15.0
seaborn==0.13.2
sentencepiece==0.2.0
shell_gpt==1.4.4
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.13.2
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tifffile==2024.9.20
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.66.5
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.44.2
triton==3.0.0
typer==0.9.4
typing==3.7.4.3
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
xxhash @ file:///work/ci_py311/python-xxhash_1676842384694/work
yarl @ file:///croot/yarl_1725976495189/work
zipp @ file:///croot/zipp_1704206909481/work
